requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-decouple==3.8

# Optional for production
gunicorn==21.2.0
whitenoise==6.6.0

# Development dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3